+++ examples/Live/liveBroadcast.php
@@ -14,6 +14,7 @@ $truncatedDebug = false;
 
 /////// MEDIA ////////
 $videoFilename = '';
+$coverPhoto = '';
 //////////////////////
 
 /////// CHARITY ID ////////
@@ -135,7 +136,8 @@ try {
 
     // Once the broadcast has ended, you can optionally add the finished
     // broadcast to your post-live feed (saved replay).
-    $ig->live->addToPostLive($broadcastId);
+    $igtvSessionId = \InstagramAPI\Signatures::generateUUID();
+    $ig->live->shareToIgtv($broadcastId, $coverPhoto, ['igtv_title' => 'test', 'igtv_session_id' => $igtvSessionId]);
 } catch (\Exception $e) {
     echo 'Something went wrong: '.$e->getMessage()."\n";
 }